Moving the UpdateStream server interface from binlog/binlogproto/ to binlog/ as the only two places using it were are already dependent on binlog/ anyway.
