+++ go/vt/binlog/binlogproto/proto.go
@@ -2,6 +2,9 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
+// Package binlogproto contains the definitions exported by the binlog
+// package, which clients can depend on, without depending on the full
+// binlog server code.
 package binlogproto
 
 // This file contains helper methods for dealing with the proto3 data
